@@ -3,7 +3,7 @@
   "version": "8.2.7",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "ng serve --host 0.0.0.0",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "rtl": "webpack --config=rtl.config.js",
@@ -70,11 +70,14 @@
     "@types/object-path": "0.11.1",
     "@types/prismjs": "1.26.0",
     "@types/sass-loader": "8.0.3",
-    "@typescript-eslint/eslint-plugin": "^8.0.0",
-    "@typescript-eslint/parser": "^8.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.12.2",
+    "@typescript-eslint/parser": "^8.12.2",
     "css-loader": "^7.1.1",
     "del": "^7.0.0",
-    "eslint": "^9.8.0",
+    "eslint": "^9.13.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsdoc": "^50.4.3",
+    "eslint-plugin-prefer-arrow": "^1.2.3",
     "jasmine-core": "~4.6.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
