@@ -1,30 +1,17 @@
-import { NgModule, APP_INITIALIZER } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HttpClientModule } from '@angular/common/http';
-import { HttpClientInMemoryWebApiModule } from 'angular-in-memory-web-api';
 import { ClipboardModule } from 'ngx-clipboard';
 import { TranslateModule } from '@ngx-translate/core';
 import { InlineSVGModule } from 'ng-inline-svg-2';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { SweetAlert2Module } from '@sweetalert2/ngx-sweetalert2';
+
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { environment } from 'src/environments/environment';
-import { SweetAlert2Module } from '@sweetalert2/ngx-sweetalert2';
-// import { AuthService } from './modules/auth/services/auth.service';
-// #fake-start#
-
-// #fake-end#
-// ====> authService: AuthService <====
-
-function appInitializer() {                                        
-  // return () => {
-  //   return new Promise((resolve) => {
-  //     //@ts-ignore
-  //     authService.getUserByToken().subscribe().add(resolve);
-  //   });
-  // };
-}
+import { CoreModule } from './core/core.module';
+import { AccountModule } from './modules/account/account.module';
 
 @NgModule({
   declarations: [AppComponent],
@@ -34,27 +21,13 @@ function appInitializer() {
     TranslateModule.forRoot(),
     HttpClientModule,
     ClipboardModule,
-    // #fake-start#
-    // environment.isMockEnabled
-    //   ? HttpClientInMemoryWebApiModule.forRoot(FakeAPIService, {
-    //     passThruUnknownUrl: true,
-    //     dataEncapsulation: false,
-    //   })
-    //   : [],
-    // #fake-end#
+    CoreModule,           // Importa o CoreModule com o AuthInterceptor configurado
     AppRoutingModule,
     InlineSVGModule.forRoot(),
     NgbModule,
     SweetAlert2Module.forRoot(),
-  ],
-  providers: [
-    // {
-    //   provide: APP_INITIALIZER,
-    //   useFactory: appInitializer,
-    //   multi: true,
-    //   // deps: [AuthService],
-    // },
+    AccountModule,
   ],
   bootstrap: [AppComponent],
 })
-export class AppModule { }
+export class AppModule {}
