@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { AuthComponent } from './auth.component';
 import { LoginComponent } from './components/login/login.component';
-import { RegistrationComponent } from './components/registration/registration.component';
+
 import { ForgotPasswordComponent } from './components/forgot-password/forgot-password.component';
 import { LogoutComponent } from './components/logout/logout.component';
 
@@ -21,10 +21,10 @@ const routes: Routes = [
         component: LoginComponent,
         data: { returnUrl: window.location.pathname },
       },
-      {
-        path: 'registration',
-        component: RegistrationComponent,
-      },
+      // {
+      //   path: 'registration',
+      //   component: RegistrationComponent,
+      // },
       {
         path: 'forgot-password',
         component: ForgotPasswordComponent,
