@@ -4,8 +4,6 @@ import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 import { AuthRoutingModule } from './auth-routing.module';
 import { LoginComponent } from './components/login/login.component';
-import { RegistrationComponent } from './components/registration/registration.component';
-import { ForgotPasswordComponent } from './components/forgot-password/forgot-password.component';
 import { LogoutComponent } from './components/logout/logout.component';
 import { AuthComponent } from './auth.component';
 import { TranslationModule } from '../i18n/translation.module';
@@ -13,8 +11,8 @@ import { TranslationModule } from '../i18n/translation.module';
 @NgModule({
   declarations: [
     LoginComponent,
-    RegistrationComponent,
-    ForgotPasswordComponent,
+    // RegistrationComponent,
+    // ForgotPasswordComponent,
     LogoutComponent,
     AuthComponent,
   ],
