@@ -24,7 +24,7 @@ export class ForgotPasswordComponent implements OnInit {
   // private fields
   private unsubscribe: Subscription[] = []; // Read more: => https://brianflove.com/2016/12/11/anguar-2-unsubscribe-observables/
   constructor(private fb: FormBuilder, private authService: AuthService) {
-    this.isLoading$ = this.authService.isLoading$;
+    // this.isLoading$ = this.authService.isLoading$;
   }
 
   ngOnInit(): void {
@@ -50,14 +50,14 @@ export class ForgotPasswordComponent implements OnInit {
     });
   }
 
-  submit() {
-    this.errorState = ErrorStates.NotSubmitted;
-    const forgotPasswordSubscr = this.authService
-      .forgotPassword(this.f.email.value)
-      .pipe(first())
-      .subscribe((result: boolean) => {
-        this.errorState = result ? ErrorStates.NoError : ErrorStates.HasError;
-      });
-    this.unsubscribe.push(forgotPasswordSubscr);
-  }
+  // submit() {
+  //   this.errorState = ErrorStates.NotSubmitted;
+  //   const forgotPasswordSubscr = this.authService
+  //     .forgotPassword(this.f.email.value)
+  //     .pipe(first())
+  //     .subscribe((result: boolean) => {
+  //       this.errorState = result ? ErrorStates.NoError : ErrorStates.HasError;
+  //     });
+  //   this.unsubscribe.push(forgotPasswordSubscr);
+  // }
 }
